add socials on about
